% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risk_quantiles.R
\name{add_risk_quantiles}
\alias{add_risk_quantiles}
\title{Add risk quantiles to a dataframe for the given column.}
\usage{
add_risk_quantiles(d, risk.col, output.col = risk.col, quants = 5,
  style = "fisher")
}
\arguments{
\item{d}{The data}

\item{risk.col}{String specifying the name of the column to quantise}

\item{output.col}{String specifying the name of the column to add (defaults to the same as risk.col)}

\item{quants}{Number of quantiles (default: 5)}

\item{style}{Method to use for calculating quantiles (passed to classIntervals; default: Fisher). One of "fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", "jenks" or "dpih"}
}
\value{
A dataframe with two columns added: one for the risk quantile (output.col_q) and one containing labels for the risk quantile (output.col_q_name)
}
\description{
Add risk quantiles to a dataframe for the given column.
}
