% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_date}
\alias{convert_date}
\title{Convert dates only when they're in Excel (numeric) format}
\usage{
convert_date(d)
}
\arguments{
\item{d}{The date}
}
\description{
Convert dates only when they're in Excel (numeric) format
}
