% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lookups.R
\name{create_lookup_lsoa_lad_county}
\alias{create_lookup_lsoa_lad_county}
\title{Create lookup table for Lower Layer Super Output Areas to Local Authorities to Counties}
\usage{
create_lookup_lsoa_lad_county(
  lookup.path = file.path(options("brclib.data_path"), "Lookups, names and codes"),
 
    lsoa.lad.name = "Lower_Layer_Super_Output_Area_2011_to_Clinical_Commissioning_Group_to_Local_Authority_District_April_2018_Lookup_in_England.csv",
 
    lad.county.name = "Local_Authority_District_to_County_December_2018_Lookup_in_England.csv"
)
}
\arguments{
\item{lookup.path}{Path to the ONS's lookup tables}

\item{lsoa.lad.name}{Name of the LSOA to LAD lookup table file (https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=name&tags=LUP_LSOA_CCG_LAD)}

\item{lad.county.name}{Name of the LAD to County lookup table file (https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=name&tags=LUP_LAD_CTY)}
}
\description{
Create lookup table for Lower Layer Super Output Areas to Local Authorities to Counties
}
