% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lookups.R
\name{create_lookup_msoa_lad_county}
\alias{create_lookup_msoa_lad_county}
\title{Create lookup table for Middle Layer Super Output Areas to Local Authorities to Counties}
\usage{

  create_lookup_msoa_lad_county(lookup.path = file.path(options("brclib.data_path"),
  "Lookups, names and codes"),
  msoa.lad.name = "Middle_Layer_Super_Output_Area_2011_to_Ward_to_LAD_December_2018_Lookup_in_England_and_Wales.csv",
  lad.county.name = "Local_Authority_District_to_County_December_2018_Lookup_in_England.csv")
}
\arguments{
\item{lookup.path}{Path to the ONS's lookup tables}

\item{msoa.lad.name}{Name of the MSOA to LAD lookup table file (https://geoportal.statistics.gov.uk/datasets/middle-layer-super-output-area-2011-to-ward-to-lad-december-2018-lookup-in-england-and-wales)}

\item{lad.county.name}{Name of the LAD to County lookup table file (https://geoportal.statistics.gov.uk/search?collection=Dataset&sort=name&tags=LUP_LAD_CTY)}
}
\description{
Create lookup table for Middle Layer Super Output Areas to Local Authorities to Counties
}
