% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maps.r
\name{download_shp}
\alias{download_shp}
\title{Download and load shapefiles from the web
- download .zip into a temp file
- unzip into a temp folder
- open and return as an `sf` object}
\usage{
download_shp(url)
}
\arguments{
\item{url}{URL to the .zip file containing shapefiles}
}
\description{
Download and load shapefiles from the web
- download .zip into a temp file
- unzip into a temp folder
- open and return as an `sf` object
}
