% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wales.R
\name{download_wales}
\alias{download_wales}
\title{Download data from Stats Wales (https://statswales.gov.wales)}
\usage{
download_wales(url)
}
\arguments{
\item{url}{The Stats Wales URL to start gathering from}
}
\description{
Download data from Stats Wales (https://statswales.gov.wales)
}
