% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maps.r
\name{get_LSOAs}
\alias{get_LSOAs}
\title{Get Lower Layer Super Output Areas and devolved nations' equivalents
- LSOA in England and Wales
- Data Zones in Scotland
- Super Output Areas in Northern Ireland}
\usage{
get_LSOAs(
 
    url_eng_wal = "https://opendata.arcgis.com/datasets/007577eeb8e34c62a1844df090a93128_0.geojson",
 
    url_sco = "http://sedsh127.sedsh.gov.uk/Atom_data/ScotGov/ZippedShapefiles/SG_DataZoneBdry_2011.zip",
 
    url_ni = "https://www.nisra.gov.uk/sites/nisra.gov.uk/files/publications/SOA2011_Esri_Shapefile_0.zip"
)
}
\arguments{
\item{url_eng_wal}{Boundaries for England and Wales (GeoJSON from ONS Geoportal)}

\item{url_sco}{Boundaries for Scotland (Shapefile)}

\item{url_ni}{Boundaries for Northern Ireland (Shapefile)}
}
\description{
Get Lower Layer Super Output Areas and devolved nations' equivalents
- LSOA in England and Wales
- Data Zones in Scotland
- Super Output Areas in Northern Ireland
}
