% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_month_num}
\alias{get_month_num}
\title{Convert month name to the month's number
(e.g. "February" or "Feb" returns 2)}
\usage{
get_month_num(month, list.of.months = month.name)
}
\arguments{
\item{month}{String containing the month, either in full (e.g. "February") or abbreviated ("Feb")}

\item{list.of.months}{Optional list of month names. Defaults to the full names (`month.name`); use `month.abb` for short month names}
}
\description{
Convert month name to the month's number
(e.g. "February" or "Feb" returns 2)
}
\examples{
get_month_num("December")
get_month_num("Dec", month.abb)

}
