% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maps.r
\name{get_MSOAs}
\alias{get_MSOAs}
\title{Middle-layer super output areas (and devolved nations' equivalents) shapefiles for
[England and Wales](https://geoportal.statistics.gov.uk/datasets/middle-layer-super-output-areas-december-2011-boundaries-ew-bsc)
[Scotland](https://www.spatialdata.gov.scot/geonetwork/srv/api/records/389787c0-697d-4824-9ca9-9ce8cb79d6f5)
 [Northern Ireland](https://www.nisra.gov.uk/publications/super-output-area-boundaries-gis-format)}
\usage{
get_MSOAs()
}
\description{
Middle-layer super output areas (and devolved nations' equivalents) shapefiles for
[England and Wales](https://geoportal.statistics.gov.uk/datasets/middle-layer-super-output-areas-december-2011-boundaries-ew-bsc)
[Scotland](https://www.spatialdata.gov.scot/geonetwork/srv/api/records/389787c0-697d-4824-9ca9-9ce8cb79d6f5)
 [Northern Ireland](https://www.nisra.gov.uk/publications/super-output-area-boundaries-gis-format)
}
