% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colours.R
\name{get_nhs_colours}
\alias{get_nhs_colours}
\title{Get a list of NHS colour hex codes
source: https://www.england.nhs.uk/nhsidentity/identity-guidelines/colours/}
\usage{
get_nhs_colours()
}
\description{
Get a list of NHS colour hex codes
source: https://www.england.nhs.uk/nhsidentity/identity-guidelines/colours/
}
