% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%**\%}
\alias{\%**\%}
\title{Multiply numbers only if at least one of them isn't zero}
\usage{
x \%**\% y
}
\arguments{
\item{x}{First number}

\item{y}{Second number}
}
\description{
Multiply numbers only if at least one of them isn't zero
}
\examples{
10 \%**\% 2 # == 20
10 \%**\% 0 # == 10

}
