% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_IMD.R
\name{load_IMD}
\alias{load_IMD}
\title{Load latest Index of Multiple Deprivation data for whole UK}
\usage{
load_IMD(
 
    url_IMD = "https://github.com/matthewgthomas/IMD/raw/master/data/UK\%20IMD\%20domains.csv"
)
}
\arguments{
\item{url_IMD}{URL for the .csv file containing all deprivation data}
}
\description{
This data file is produced by: https://github.com/matthewgthomas/IMD
}
