% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{load_postcodes}
\alias{load_postcodes}
\title{Load BRC's bespoke postcode directory}
\usage{
load_postcodes(postcode.path = file.path(options("brclib.data_path"),
  "Postcodes"),
  postcode.name = "National_Statistics_Postcode_Lookup - BRC.csv")
}
\arguments{
\item{postcode.path}{Where the postcodes file is stored}

\item{postcode.name}{Name of the postcodes file}
}
\description{
Load BRC's bespoke postcode directory
}
