% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodes.R
\name{postcode_regex}
\alias{postcode_regex}
\title{Regular expression to match postcodes (allowing lowercase and unlimited spaces)
source: https://stackoverflow.com/a/7259020
see also: page 6 of https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/488478/Bulk_Data_Transfer_-_additional_validation_valid_from_12_November_2015.pdf}
\usage{
postcode_regex()
}
\description{
Regular expression to match postcodes (allowing lowercase and unlimited spaces)
source: https://stackoverflow.com/a/7259020
see also: page 6 of https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/488478/Bulk_Data_Transfer_-_additional_validation_valid_from_12_November_2015.pdf
}
