% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brc_people.R
\name{process_CRVs}
\alias{process_CRVs}
\title{Clean and process Community Reserve Volunteers data}
\usage{
process_CRVs(vols.dir = file.path(options("brclib.data_path"),
  "Volunteers"), vols.file.latest = "CRVs - latest.csv")
}
\arguments{
\item{vols.dir}{Path to volunteers data}

\item{vols.file.latest}{What to call the output file}
}
\description{
Clean and process Community Reserve Volunteers data
}
