% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_entirely.R
\name{read_excel_entirely}
\alias{read_excel_entirely}
\title{Load all worksheets in an Excel file into a single dataframe}
\usage{
read_excel_entirely(path)
}
\arguments{
\item{path}{The Excel file to read}
}
\description{
Load all worksheets in an Excel file into a single dataframe
}
