% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{round_up}
\alias{round_up}
\title{Round up to nearest specified number
source: https://stackoverflow.com/a/6468532}
\usage{
round_up(x, to = 10)
}
\arguments{
\item{x}{Number to round up}

\item{to}{Number to round to to}
}
\description{
Round up to nearest specified number
source: https://stackoverflow.com/a/6468532
}
