% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{round_up_nice}
\alias{round_up_nice}
\title{Round up to nearest 'nice' number
source: https://stackoverflow.com/a/6463946}
\usage{
round_up_nice(x, nice = c(1, 2, 4, 5, 6, 8, 10))
}
\arguments{
\item{x}{Number to round up}

\item{nice}{Vector of 'nice' numbers (no need to change this)}
}
\description{
Round up to nearest 'nice' number
source: https://stackoverflow.com/a/6463946
}
