% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_risk_quantiles.R
\name{add_risk_quantiles}
\alias{add_risk_quantiles}
\title{Add risk quantiles to a dataframe for the given column.}
\usage{
add_risk_quantiles(
  d,
  risk.col,
  output.col = risk.col,
  quants = 5,
  highest.number.is.worst = TRUE,
  style = "fisher",
  samp_prop = 1
)
}
\arguments{
\item{d}{The data}

\item{risk.col}{String specifying the name of the column to quantise}

\item{output.col}{String specifying the name of the column to add (defaults to the same as risk.col)}

\item{quants}{Number of quantiles (default: 5)}

\item{highest.number.is.worst}{Should a risk score of 1 represent the highest number in the data (TRUE) or the lowest (FALSE)?}

\item{style}{Method to use for calculating quantiles (passed to classIntervals; default: Fisher). One of "fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", "jenks" or "dpih"}

\item{samp_prop}{The proportion of samples to use, if slicing using "fisher" or "jenks" (passed to classIntervals; default: 100%)}
}
\value{
A dataframe with two columns added: one for the risk quantile (output.col_q) and one containing labels for the risk quantile (output.col_q_name)
}
\description{
Add risk quantiles to a dataframe for the given column.
}
